backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowva-backend"
version = "0.1.0"
description = "Knowva - AI Reading Experience Support Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-cloud-firestore>=2.16.0",
    "firebase-admin>=6.5.0",
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/knowva"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
